odop/requirements.txt
cloudpickle==3.0.0
fastapi==0.110.0
flatten_dict==0.4.2
numpy==1.24.1
psutil==5.9.0
pydantic==1.10.4
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
scipy==1.12.0
tabulate==0.9.0
tensorflow==2.15.0.post1
tinyflux==0.4.1
